исправления после проверки
