+++ source/js/script.js
@@ -8,7 +8,7 @@ import result from './modules/result.js';
 import form from './modules/form.js';
 import social from './modules/social.js';
 import FullPageScroll from './modules/full-page-scroll';
-import custom from './modules/custom'
+import custom from './modules/custom';
 
 // init modules
 mobileHeight();
