+++ source/scss/blocks/custom.scss
@@ -500,3 +500,8 @@
   margin-top: -3rem;
   height: 22.6rem;
 }
+
+
+.game__counter {
+  width: 11rem;
+}
